Extract Blacklight::Solr::SearchBuilder with processing chain for converting blacklight parameters to solr request parameters
