+++ lib/blacklight/solr.rb
@@ -5,5 +5,6 @@ module Blacklight::Solr
   autoload :FacetPaginator, 'blacklight/solr/facet_paginator'
   autoload :Document, 'blacklight/solr/document'
   autoload :Request, 'blacklight/solr/request'
+  autoload :SearchBuilder, 'blacklight/solr/search_builder'
   
 end
