+++ spec/lib/blacklight/search_helper_spec.rb
@@ -59,213 +59,10 @@ describe Blacklight::SearchHelper do
         solr_params[:wt] = "TESTING"
       end
 
-      expect(Deprecation).to receive(:warn)
+      allow(Deprecation).to receive(:warn)
       expect(subject.solr_search_params({}, [:add_foo_to_solr_params])[:wt]).to eq "TESTING"
     end
   end
-
-  describe "facet_value_to_fq_string" do
-
-    it "should use the raw handler for strings" do
-      expect(subject.send(:facet_value_to_fq_string, "facet_name", "my value")).to eq "{!raw f=facet_name}my value"
-    end
-
-    it "should pass booleans through" do
-      expect(subject.send(:facet_value_to_fq_string, "facet_name", true)).to eq "facet_name:true"
-    end
-
-    it "should pass boolean-like strings through" do
-      expect(subject.send(:facet_value_to_fq_string, "facet_name", "true")).to eq "facet_name:true"
-    end
-
-    it "should pass integers through" do
-      expect(subject.send(:facet_value_to_fq_string, "facet_name", 1)).to eq "facet_name:1"
-    end
-
-    it "should pass integer-like strings through" do
-      expect(subject.send(:facet_value_to_fq_string, "facet_name", "1")).to eq "facet_name:1"
-    end
-
-    it "should pass floats through" do
-      expect(subject.send(:facet_value_to_fq_string, "facet_name", 1.11)).to eq "facet_name:1\\.11"
-    end
-
-    it "should pass floats through" do
-      expect(subject.send(:facet_value_to_fq_string, "facet_name", "1.11")).to eq "facet_name:1\\.11"
-    end
-
-    it "should escape negative integers" do
-      expect(subject.send(:facet_value_to_fq_string, "facet_name", -1)).to eq "facet_name:\\-1"
-    end
-
-    it "should pass date-type fields through" do
-      allow(blacklight_config.facet_fields).to receive(:[]).with('facet_name').and_return(double(:date => true, :query => nil, :tag => nil))
-
-      expect(subject.send(:facet_value_to_fq_string, "facet_name", "2012-01-01")).to eq "facet_name:2012\\-01\\-01"
-    end
-
-    it "should escape datetime-type fields" do
-      allow(blacklight_config.facet_fields).to receive(:[]).with('facet_name').and_return(double(:date => true, :query => nil, :tag => nil))
-
-      expect(subject.send(:facet_value_to_fq_string, "facet_name", "2003-04-09T00:00:00Z")).to eq "facet_name:2003\\-04\\-09T00\\:00\\:00Z"
-    end
-    
-    it "should format Date objects correctly" do
-      allow(blacklight_config.facet_fields).to receive(:[]).with('facet_name').and_return(double(:date => nil, :query => nil, :tag => nil))
-      d = DateTime.parse("2003-04-09T00:00:00")
-      expect(subject.send(:facet_value_to_fq_string, "facet_name", d)).to eq "facet_name:2003\\-04\\-09T00\\:00\\:00Z"      
-    end
-
-    it "should handle range requests" do
-      expect(subject.send(:facet_value_to_fq_string, "facet_name", 1..5)).to eq "facet_name:[1 TO 5]"
-    end
-
-    it "should add tag local parameters" do
-      allow(blacklight_config.facet_fields).to receive(:[]).with('facet_name').and_return(double(:query => nil, :tag => 'asdf', :date => nil))
-
-      expect(subject.send(:facet_value_to_fq_string, "facet_name", true)).to eq "{!tag=asdf}facet_name:true"
-      expect(subject.send(:facet_value_to_fq_string, "facet_name", "my value")).to eq "{!raw f=facet_name tag=asdf}my value"
-    end
-  end
-
-  describe "converts a String fq into an Array" do
-    it "should return the correct overriden parameter" do
-      solr_parameters = {:fq => 'a string' }
-
-      subject.add_facet_fq_to_solr(solr_parameters, {})
-
-      expect(solr_parameters[:fq]).to be_a_kind_of Array
-    end
-  end
-
-  describe "#add_solr_fields_to_query" do
-    let(:blacklight_config) do
-      config = Blacklight::Configuration.new do |config|
-
-        config.add_index_field 'an_index_field', solr_params: { 'hl.alternativeField' => 'field_x'}
-        config.add_show_field 'a_show_field', solr_params: { 'hl.alternativeField' => 'field_y'}
-        config.add_field_configuration_to_solr_request!
-      end
-    end
-
-    let(:solr_parameters) do
-      solr_parameters = Blacklight::Solr::Request.new
-
-      subject.add_solr_fields_to_query(solr_parameters, {})
-
-      solr_parameters
-    end
-
-    it "should add any extra solr parameters from index and show fields" do
-      expect(solr_parameters[:'f.an_index_field.hl.alternativeField']).to eq "field_x"
-      expect(solr_parameters[:'f.a_show_field.hl.alternativeField']).to eq "field_y"
-    end
-  end
-
-  describe "#add_facetting_to_solr" do
-
-    let(:blacklight_config) do
-       config = Blacklight::Configuration.new
-
-       config.add_facet_field 'test_field', :sort => 'count'
-       config.add_facet_field 'some-query', :query => {'x' => {:fq => 'some:query' }}, :ex => 'xyz'
-       config.add_facet_field 'some-pivot', :pivot => ['a','b'], :ex => 'xyz'
-       config.add_facet_field 'some-field', solr_params: { 'facet.mincount' => 15 }
-       config.add_facet_fields_to_solr_request!
-
-       config
-    end
-
-    let(:solr_parameters) do
-      solr_parameters = Blacklight::Solr::Request.new
-      
-      subject.add_facetting_to_solr(solr_parameters, {})
-
-      solr_parameters
-    end
-
-    it "should add sort parameters" do
-      expect(solr_parameters[:facet]).to be true
-
-      expect(solr_parameters[:'facet.field']).to include('test_field')
-      expect(solr_parameters[:'f.test_field.facet.sort']).to eq 'count'
-    end
-
-    it "should add facet exclusions" do
-      expect(solr_parameters[:'facet.query']).to include('{!ex=xyz}some:query')
-      expect(solr_parameters[:'facet.pivot']).to include('{!ex=xyz}a,b')
-    end
-
-    it "should add any additional solr_params" do
-      expect(solr_parameters[:'f.some-field.facet.mincount']).to eq 15
-    end
-
-    describe ":include_in_request" do
-      let(:solr_parameters) do
-        solr_parameters = Blacklight::Solr::Request.new
-        subject.add_facetting_to_solr(solr_parameters, {})
-        solr_parameters
-      end
-
-      it "should respect the include_in_request parameter" do
-        blacklight_config.add_facet_field 'yes_facet', include_in_request: true
-        blacklight_config.add_facet_field 'no_facet', include_in_request: false
-        
-        expect(solr_parameters[:'facet.field']).to include('yes_facet')
-        expect(solr_parameters[:'facet.field']).not_to include('no_facet')
-      end
-
-      it "should default to including facets if add_facet_fields_to_solr_request! was called" do
-        blacklight_config.add_facet_field 'yes_facet'
-        blacklight_config.add_facet_field 'no_facet', include_in_request: false
-        blacklight_config.add_facet_fields_to_solr_request!
-
-        expect(solr_parameters[:'facet.field']).to include('yes_facet')
-        expect(solr_parameters[:'facet.field']).not_to include('no_facet')
-      end
-    end
-
-    describe ":add_facet_fields_to_solr_request!" do
-
-      let(:blacklight_config) do
-        config = Blacklight::Configuration.new
-        config.add_facet_field 'yes_facet', include_in_request: true
-        config.add_facet_field 'no_facet', include_in_request: false
-        config.add_facet_field 'maybe_facet'
-        config.add_facet_field 'another_facet'
-        config
-      end
-
-      let(:solr_parameters) do
-        solr_parameters = Blacklight::Solr::Request.new
-        subject.add_facetting_to_solr(solr_parameters, {})
-        solr_parameters
-      end
-
-      it "should add facets to the solr request" do
-        blacklight_config.add_facet_fields_to_solr_request!
-        expect(solr_parameters[:'facet.field']).to match_array ['yes_facet', 'maybe_facet', 'another_facet']
-      end
-
-      it "should not override field-specific configuration by default" do
-        blacklight_config.add_facet_fields_to_solr_request!
-        expect(solr_parameters[:'facet.field']).to_not include 'no_facet'
-      end
-
-      it "should allow white-listing facets" do
-        blacklight_config.add_facet_fields_to_solr_request! 'maybe_facet'
-        expect(solr_parameters[:'facet.field']).to include 'maybe_facet'
-        expect(solr_parameters[:'facet.field']).not_to include 'another_facet'
-      end
-
-      it "should allow white-listed facets to override any field-specific include_in_request configuration" do
-        blacklight_config.add_facet_fields_to_solr_request! 'no_facet'
-        expect(solr_parameters[:'facet.field']).to include 'no_facet'
-      end
-    end
-  end
-
-
 
   describe "solr_facet_params" do
     before do
@@ -339,22 +136,6 @@ describe Blacklight::SearchHelper do
     end
   end
 
-  describe "for facet limit parameters config ed" do                
-    # subject do
-    #   subject.solr_search_params({ search_field: "test_field", :q => "test query"}, {}, default_method_chain)
-    # end
-
-    # let(:blacklight_config) { copy_of_catalog_config }
-    # 
-    # it "should include specifically configged facet limits +1" do
-    #   expect(subject[:"f.subject_topic_facet.facet.limit"]).to eq 21      
-    # end
-    # it "should not include a facet limit for a nil key in hash" do
-    #   expect(subject).not_to have_key(:"f.format.facet.limit")
-    #   expect(subject).not_to have_key(:"facet.limit")
-    # end
-  end
-  
   describe "get_facet_pagination", :integration => true do
     before do
       Deprecation.silence(Blacklight::SearchHelper) do
@@ -947,19 +728,6 @@ describe Blacklight::SearchHelper do
       blacklight_config.index.group = 'xyz'
       expect(subject.grouped_key_for_results).to eq('xyz')
     end 
-  end
-
-  describe "#with_tag_ex" do
-    it "should add an !ex local parameter if the facet configuration requests it" do
-      expect(subject.with_ex_local_param("xyz", "some-value")).to eq "{!ex=xyz}some-value"
-    end
-
-    it "should not add an !ex local parameter if it isn't configured" do
-      mock_field = double()
-      expect(subject.with_ex_local_param(nil, "some-value")).to eq "some-value"
-    end
-
-
   end
 
   describe "#get_previous_and_next_documents_for_search" do
